fix integrity running before db migrations have applied
